Account UI and manager

+++ app/managers/account_manager.py
@@ -1,5 +1,5 @@
 import hashlib
-from datetime import date, timedelta
+from datetime import datetime, date, timedelta
 
 class AccountManager:
     def __init__(self, db):
@@ -11,10 +11,16 @@ class AccountManager:
         self.metric_value_table = self.dbm.models["MetricValue"]
         self.challenges_table = self.dbm.models["Challenges"]
         self.user_challenges_table = self.dbm.models["UserChallenges"]
+        self.user_reward_table = self.dbm.models["UserRewards"]
+        self.user_task_table = self.dbm.models["UserTask"]
+        self.user_routine_table = self.dbm.models["UserRoutine"]
+        self.friends_table = self.dbm.models["Friends"]
+        self.comp_participant_table = self.dbm.models["CompParticipant"]
+
 
     def _validate_username(self, username):
-        if not username or len(username) < 3:
-            raise ValueError("Username must be at least 3 characters long")
+        if not username or len(username) > 30 or len(username) < 5:
+            raise ValueError("Username must be at between 5 and 30 characters long")
         if not username.isalnum():
             raise ValueError("Username can only contain letters and numbers")
         # Check that username has at least 1 letter in it
@@ -30,12 +36,26 @@ class AccountManager:
         if "@" not in email or "." not in email.split("@")[1]:
             raise ValueError("Invalid email format")
         return True
+    
+    def _validate_fname(self, fname):
+        if not fname or len(fname) > 20 or len(fname) < 3:
+            raise ValueError("First name must be at between 3 and 20 characters long")
+        if not fname.isalpha():
+            raise ValueError("First name can only contain letters")
+        return True
+    
+    def _validate_dob(self, dob):
+        try:
+            datetime.strptime(dob, "%Y-%m-%d")
+            return True
+        except ValueError:
+            raise ValueError("Date must be in valid format")
 
     def create_account(self, username, password, fname, email, dob):
-        if self.user_info(username):
+        if self.user_info(username=username):
             return {'success': False, 'message': "Account creation failed: Username already exists"}
         try:
-            if self._validate_username(username) and self._validate_email(email):
+            if self._validate_username(username) and self._validate_email(email) and self._validate_fname(fname) and self._validate_dob(dob):
                 user = self.dbm.create_record(
                     self.users_table,
                     username=username,
@@ -73,12 +93,16 @@ class AccountManager:
                 chall_edate=(date.today() + timedelta(days=7))
             )
 
-    def user_info(self, username):
+    def user_info(self, username=None, user_id=None):
         try: 
-            user = self.users_table.get(self.users_table.username == username)
+            if user_id:
+                user = self.users_table.get(self.users_table.user_id == user_id)
+            else:
+                user = self.users_table.get(self.users_table.username == username)
         except Exception:
             return None
         return {
+            'success': True,
             'user id': user.user_id,
             'username': user.username,
             'email': user.user_email,
@@ -88,25 +112,56 @@ class AccountManager:
             } 
 
     def delete_account(self, user_id):
-        user = self.dbm.delete_record(self.users_table, user_id)
+        # Check user exists
+        if self.dbm.read_record(self.users_table, user_id):
+            # First delete all records associated with user_id
+            self.metric_value_table.delete().where(self.metric_value_table.user_id == user_id).execute()
+            self.user_challenges_table.delete().where(self.user_challenges_table.user_id == user_id).execute()
+            self.user_reward_table.delete().where(self.user_reward_table.user_id == user_id).execute()
+            self.user_task_table.delete().where(self.user_task_table.user_id == user_id).execute()
+            self.user_routine_table.delete().where(self.user_routine_table.user_id == user_id).execute()
+            self.friends_table.delete().where(self.friends_table.user_id == user_id).execute()
+            self.comp_participant_table.delete().where(self.comp_participant_table.user_id == user_id).execute()
+
+            # Then delete the user
+            user = self.dbm.delete_record(self.users_table, user_id)
+            if user == 1:
+                return {'success': True, 'message': f"User ID {user_id} was deleted"}
+            elif user == 0:
+                return {'success': False, 'message': f"Failed to delete user ID {user_id}"}
+        return {'success': False, 'message': f"User ID {user_id} does not exist"}
+
+    def update_username(self, user_id, new_username):
+        try:
+            if self._validate_username(new_username):
+                user = self.dbm.update_record(self.users_table, user_id, username=new_username)
+        except Exception as e:
+            return {'success': False, 'message': f"Username update failed: {e}"}
         if user is None:
-            print("This User Doesn't Exist")
-        
+            return {'success': False, 'message': f"User ID {user_id} does not exist"}
+        return {'success': True, 'message': "Username updated"}
+
     def update_email(self, user_id, new_email):
-        user = self.dbm.update_record(self.users_table, user_id, user_email=new_email)
+        try:
+            if self._validate_email(new_email):
+                user = self.dbm.update_record(self.users_table, user_id, user_email=new_email)
+        except Exception as e:
+            return {'success': False, 'message': f"Email update failed: {e}"}
         if user is None:
-            print("This User Doesn't Exist")
-            return
+            return {'success': False, 'message': f"User ID {user_id} does not exist"}
+        return {'success': True, 'message': f"Email updated"}
    
     def update_password(self, user_id, new_password):
-        new_password = hashlib.sha256(new_password.encode('utf-8')).hexdigest()
-        user = self.dbm.update_record(self.users_table, user_id, user_password=new_password)
+        try:
+            user = self.dbm.update_record(self.users_table, user_id, user_password=new_password)
+        except Exception as e:
+            return {'success': False, 'message': f"Password update failed: {e}"}
         if user is None:
-            print("This User Doesn't Exist")
-            return
+            return {'success': False, 'message': f"User ID {user_id} does not exist"}
+        return {'success': True, 'message': "Password updated"}
     
     def login(self, username, password):
-        user = self.user_info(username)
+        user = self.user_info(username=username)
         # If user exists, check password
         if user:
             if user['password'] == password:
